+++ ambientika-local-control/src/services/ha-auto-discovery.service.ts
@@ -136,12 +136,6 @@ export class HAAutoDiscoveryService {
         return this.getSensorDiscoveryMessage('preset_mode', process.env.PRESET_MODE_STATE_TOPIC, device, attributes);
     }
 
-    getHouseIdSensorMessage(device: Device): string {
-        const attributes: HaAutoDiscoverDeviceAttributes = {
-            icon: 'mdi:home-account'
-        };
-        return this.getSensorDiscoveryMessage('house_id', process.env.HOUSE_ID_TOPIC, device, attributes);
-    }
 
     private getBinarySensorDiscoveryMessage(type: string, topic: string | undefined, device: Device,
                                             attributes?: HaAutoDiscoverDeviceAttributes): string {